Refactor jinjax
