+++ src/fastlife/adapters/jinjax/jinjax_ext/__init__.py
@@ -1,0 +1,1 @@
+"""JinjaX extention to support introspection."""
