+++ tests/unittests/templates/components/SyntaxError.jinja
@@ -1,0 +1,8 @@
+{# def
+  class: Annotated[
+    str,
+    "Unfortunatly class is a reserved word in Python and not handled by JinjaX"
+  ] = ""
+#}
+<div class="{{class}}">
+</div>
